backend/requirements.txt
fastapi==0.115.14
uvicorn==0.34.3
motor==3.7.1
python-dotenv==1.0.0
pydantic==2.11.7
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
